move specialChars to jest-util
