+++ packages/jest-cli/src/constants.js
@@ -7,15 +7,5 @@
  * @flow
  */
 
-const isWindows = process.platform === 'win32';
-
-export const CLEAR = isWindows ? '\x1B[2J\x1B[0f' : '\x1B[2J\x1B[3J\x1B[H';
-export const ARROW = ' \u203A ';
-export const ICONS = {
-  failed: isWindows ? '\u00D7' : '\u2715',
-  pending: '\u25CB',
-  success: isWindows ? '\u221A' : '\u2713',
-  todo: '\u270E',
-};
 export const PACKAGE_JSON = 'package.json';
 export const JEST_CONFIG = 'jest.config.js';
